requirements.txt
anthropic>=0.39.0
openai>=1.50.0
google-genai>=1.0.0
pyyaml>=6.0
rich>=13.0.0
